refactor: update import paths for components to align with new directory structure
